The test directories have been renamed with more sensible names.
I have updated the model.e file to produce a more correct representation of
viscoe-elasticity. The boundary conditions now work much better and is
equivalent with PML.

The ac2d.e file has been updated with visco-elastic bulk modulues and
visco-elastic density. All equations of motion should now be correct.

The notes.tex file in the Doc directory contains a (very unfinished) list
of equations for a general elastic 3D medium, and then
reduced to the 2D acoustic case.
Also the most important references are given in the notes.